connection to mysql database, sample greetings storage and retrieval
Default MySQL Host: localhost
Can be configured via --mysql-host Parameter. For example, if you run your MySQL server in Docker on Windows then you would start the application with --mysql-host=192.168.99.100

The GreetingController now accepts new Greetings as as POST request, you can view the greetings you have received with a GET request.
The storage is done by GreetingRepository, which is an interface. Spring parses the method names and deduces that the findByName method needs an implementation that searches the database for greetings with the given name and generates such an implementation. You will most likely not need to write custom SQL code, as Spring is able to generate that for you.
Also contained is a docker-compose file that starts a MySQL Database on Port 3306 with the database and user pregenerated. This is useful for development if you don't want to setup a MySQL server on your development machine.